lib/util: Got config_parser to compile with GNU 4.8.3
I found that I needed to include <unistd.h> to get
access to the sleep() function.

+++ lib/util/source/config_parser.cxx
@@ -11,6 +11,7 @@
 #include <vector>
 #include <map>
 #include <stdexcept>
+#include <unistd.h>
 #include "config_parser.hh"
 using namespace std;
 
@@ -22,7 +23,7 @@ ConfigParser::ConfigParser( string fname )
     // and setting up the config_map
     ifstream infile( fname.c_str() );    
     size_t retries = 10;
-    for ( int i=0; i< retries && infile.fail(); ++i ) {
+    for ( size_t i=0; i< retries && infile.fail(); ++i ) {
         infile.open( fname.c_str() );
         sleep(2);
     }
